Fixed linting errors
