+++ .eslintignore
@@ -1,2 +1,8 @@
 # ignore inner node_modules from script
-**/node_modules/**
+/src/scripts/node_modules/**
+
+# this looks to be some kind of boilerplate
+/src/serviceWorker.js
+
+# I have no idea what this is
+/src/App.test.js
