+++ src/components/SocialsBar.jsx
@@ -1,21 +1,19 @@
 import React from 'react';
 import './Socials.css';
 
-function SocialsBar(props) {
-    const { flow } = props;
-
+function SocialsBar({ flow, socials }) {
     return (
         <div className="socialsContainer" style={{ flexDirection: flow }}>
-            {props.socials.map((social) => (
-            <a href={social.link} target="_blank" rel="noopener noreferrer">
-                  <div className={`${social.type}Container socContainer`}>
+            {socials.map((social) => (
+                <a href={social.link} target="_blank" rel="noopener noreferrer">
+                    <div className={`${social.type}Container socContainer`}>
 
                         <i className={`fab fa-${social.type}${social.type === 'facebook' ? '-f' : ''}`} />
 
                     </div>
                 </a>
             ))}
-      </div>
+        </div>
 
     );
 }
