--- src/components/data/Data.js
@@ -1,52 +1,0 @@
-import React from 'react';
-
-// export const homepageVideoURL = "https://cdn-cf-east.streamable.com/video/mp4/wduiok.mp4?Expires=1613007120&Signature=JfEttG1i1xWVA7Up6UbrPcNwSQvW5mkbwMRq~BHZ~~JvdU~AAXZz9f6gAzNfu9kqfxGJnd8IlsGDcssi13UDmLw3-VJ3dzqlyPg7vcDe8d291JwclGcuC9eSU4T5S3~3Kr4I8MkNvYsYeeir8KKzdoCJrXqV2Qv6DL~I1IKNpe3Ov~dbeWmE72LqjhP1Fq1qCu4YtHj7HrRseGLvVMgFaPFvtYxXSEsNy6ZAh3fg9xZwA4fuYEPyYmDOSydIS~xKMdujU4zP3xW6DGNbI0GjQpLqypIyRL4XAlkyaTMZDaJ9mYCMYrb~S8o2nHVjASvcGj6M90MZc6vwG6OfRzUD~Q__&Key-Pair-Id=APKAIEYUVEN4EVB2OKEQ";
-export const homepageVideoURL = '/Videos/homepage-video.mp4';
-
-export const heroData = {
-    headline: 'Rutgers Esports',
-    slogan: 'Live, Love, Laugh',
-    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc ut ultrices justo. Quisque euismod sed lorem et convallis. Praesent finibus turpis elit. Aenean varius suscipit libero, nec finibus mauris tristique eu.',
-    videoSourceURL: homepageVideoURL,
-};
-
-// Basic Page Data
-export const competitivePage = {
-    page_title: 'Competitive',
-    page_content: <span>
-        <p>
-        No Esports organization would be complete without competition, and Rutgers has a number of powerful competitive teams. We play throughout the semester for fun and glory, and we’re always looking for new people to spar against or potentially join us. On top of that, we also have connections to major teams like New York Excelsior, as well as alumni who have gone on to work in competitive Esports.
-        </p>
-        <b>Teams</b>
-        <p>
-        Rutgers University has teams for Rocket reague, DOTA 2, CS:GO, and more. Tryouts for each happen every semester, and every Rutgers student is welcome to try and join. We compete in collegiate tournaments against other schools, and most of our games are streamed on twitch.tv/rutgersesports. Most teams have both an A-team and a B-team, generally designated Division I and Division II. 
-        </p>
-        <b>Coaching</b>
-        <p>
-        Even if you didn’t make a competitive team, we still have a number of resources for you to get better. The Rutgers Esports discord server is home to a number of coaches for each game and some are willing to do VOD reviews or even give one-on-one practice. It’s never too late to improve, and our coaches would love to help you.
-        </p>
-        <b>Tournaments</b>
-        <p>
-        Don’t have enough time to devote to a team, but still want to compete? We have a number of tournaments throughout the semester, so join our discord so you can find out when and how to compete. Some games have more tournaments than others, Smash Ultimate competes biweekly at their meetings, while others like Overwatch only have one or two a semester, so get your friends together and prepare for some of the most exciting challenges of your career.
-        </p>
-    </span>,
-};
-
-export const communityPage = {
-    page_title: 'Our Community',
-    page_content: <span>
-        <p>Rutgers Esports is one of the largest Esports Organizations in the country, with hundreds of active members and a number of alumni involved in Esports. We strive to build a healthy, inclusive community where everyone is welcome to come out and play games.</p>
-        <b>Weekly Meetings</b>
-        <p>
-        Every Friday of the semester from 8 PM to Midnight, we host our club meetings on the second floor of the SERC. Each room is occupied by a club like Smash Ultimate, Overwatch, League of Legends, and many more. Here, we play games, watch VODs, host tournaments, and usually get something to eat. Our meetings are a great place to meet friends who are into the same games as you are, as well as test your skills against some of the best gamers in school. You can come and go as you please, so come on out to say hi.
-        </p>
-        <b>Events</b>
-        <p>
-        Rutgers Esports throws some of the biggest and greatest collegiate gaming events, and they’re all right here on campus. We kick off every semester with a smaller, friendly event where we have a few tournaments, giveaways, and an event-long stream. This is a great place to get to know the community and find some games to play. Later on in the semester is our big event: the Fireside Open in Fall and the Scarlet Classic in Spring.
-        </p>
-        <b>Discord</b>
-        <p>
-        On top of all that, we also have a Discord server. We keep you posted on upcoming meetings and events, there’s a school survival channel where you can discuss assignments with fellow students, and of course a general chat to hang out in and get to know people. Any in-houses are ran through the discord, and we occasionally host larger events and games, so come join us at discord.gg/rutgersesports
-        </p>
-    </span>,
-};
