+++ src/index.jsx
@@ -4,8 +4,8 @@ import './index.css';
 import App from './App';
 
 ReactDOM.render(
-  <React.StrictMode>
-      <App />
+    <React.StrictMode>
+        <App />
     </React.StrictMode>,
     document.getElementById('root'),
 );
