+++ src/scripts/download-large-files.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 const fs = require('fs');
 const ytdl = require('ytdl-core');
 
@@ -8,14 +9,14 @@ console.log('Started download-large-files.js script.');
 // videos
 if (sources.videos) {
     const { videos } = sources;
-    for (const video of videos) {
+    videos.forEach((video) => {
         const { filename, link } = video;
         const filePath = `../../public/Videos/${filename}`;
         const doesFileExist = fs.existsSync(filePath);
 
         if (!doesFileExist) { console.log(`Started downloading ${filename}.`); } else {
             console.log(`${filename} exists. Not redownloading. Delete this file locally and rerun the script to download.`);
-            continue;
+            return;
         }
 
         const writeableStream = ytdl(link)
@@ -28,7 +29,7 @@ if (sources.videos) {
         writeableStream.on('error', () => {
             console.log(`Error downloading ${filename}.`);
         });
-    }
+    });
 }
 
 // ... other types of large files
